feat:  add remote build support - extend host_backend client with remote build support - refactored `cli.py` for better clarity - update tests
