+++ libs/cli/tests/unit_tests/cli/test_cli.py
@@ -9,7 +9,7 @@ from pathlib import Path
 
 from click.testing import CliRunner
 
-import langgraph_cli.cli as cli_module
+import langgraph_cli.deploy as deploy_module
 from langgraph_cli.cli import cli, prepare_args_and_stdin
 from langgraph_cli.config import Config, _get_pip_cleanup_lines, validate_config
 from langgraph_cli.docker import DEFAULT_POSTGRES_URI, DockerCapabilities, Version
@@ -344,7 +344,7 @@ def test_deploy_list_command(monkeypatch) -> None:
                 ]
             }
 
-    monkeypatch.setattr(cli_module, "HostBackendClient", FakeClient)
+    monkeypatch.setattr(deploy_module, "HostBackendClient", FakeClient)
 
     result = runner.invoke(
         cli,
@@ -384,7 +384,7 @@ def test_deploy_list_command_no_results(monkeypatch) -> None:
         def list_deployments(self, name_contains: str = ""):
             return {"resources": []}
 
-    monkeypatch.setattr(cli_module, "HostBackendClient", FakeClient)
+    monkeypatch.setattr(deploy_module, "HostBackendClient", FakeClient)
 
     result = runner.invoke(
         cli,
@@ -416,7 +416,7 @@ def test_deploy_delete_command(monkeypatch) -> None:
             captured["deployment_id"] = deployment_id
             return None
 
-    monkeypatch.setattr(cli_module, "HostBackendClient", FakeClient)
+    monkeypatch.setattr(deploy_module, "HostBackendClient", FakeClient)
 
     result = runner.invoke(
         cli,
@@ -458,7 +458,7 @@ def test_deploy_delete_command_cancelled(monkeypatch) -> None:
             deleted = True
             return None
 
-    monkeypatch.setattr(cli_module, "HostBackendClient", FakeClient)
+    monkeypatch.setattr(deploy_module, "HostBackendClient", FakeClient)
 
     result = runner.invoke(
         cli,
@@ -493,7 +493,7 @@ def test_deploy_delete_command_force(monkeypatch) -> None:
             captured["deployment_id"] = deployment_id
             return None
 
-    monkeypatch.setattr(cli_module, "HostBackendClient", FakeClient)
+    monkeypatch.setattr(deploy_module, "HostBackendClient", FakeClient)
 
     result = runner.invoke(
         cli,
